pyproject.toml
[tool.poetry]
name = "the-dark-closet"
version = "0.1.0"
description = "A Pygame adventure where a boy conquers a fantasy world."
authors = ["The Dark Closet Team <maintainers@example.com>"]
readme = "README.md"
packages = [{ include = "the_dark_closet", from = "src" }]

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
pygame-ce = "^2.5"

[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
ruff = "^0.6.9"
mypy = "^1.11.2"
pylint = "^3.2.7"
types-setuptools = "^75.1.0.20241025"
pytest = "^8.0.0"
pytest-cov = "^4.0.0"
pytest-html = "^4.0.0"
pytest-xdist = "^3.0.0"

[tool.poetry.scripts]
# Run with: poetry run the-dark-closet
the-dark-closet = "the_dark_closet.__main__:main"
tdc-tests = "scripts.run_action_tests:main"

[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true
warn_redundant_casts = true
warn_unused_ignores = true
show_error_codes = true
pretty = true

[tool.ruff]
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F"]
extend-ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
"scripts/*.py" = ["E402"]

[tool.pylint.MASTER]
ignore = [".venv", "build", "dist"]
jobs = 0

[tool.pylint.'MESSAGES CONTROL']
disable = [
  "missing-module-docstring",
  "missing-class-docstring",
  "missing-function-docstring",
  "no-member",  # pygame constants not recognized
  "line-too-long",  # allow longer lines
  "too-many-instance-attributes",
  "too-many-branches",
  "too-many-locals",
  "too-many-statements",
  "too-few-public-methods",
  "broad-exception-caught",
  "cell-var-from-loop",
  "no-else-return",
  "wrong-import-position",
  "import-outside-toplevel",
  "invalid-name",
  "unused-import",
  "too-many-arguments",
  "too-many-positional-arguments",
  "reimported",
  "unused-variable",
]

[tool.pylint.BASIC]
good-names = [
  "dx", "dy", "x", "y", "tx", "ty",
]
